+++ src/mkvtoolnix-gui/merge/input.cpp
@@ -1572,7 +1572,10 @@ Tab::setOutputFileNameMaybe(bool force) {
   else
     Q_ASSERT_X(false, "setOutputFileNameMaybe", "Untested destination file name policy");
 
-  auto baseName              = m_config.m_title.isEmpty() ? QFileInfo{ m_config.m_firstInputFileName }.completeBaseName() : m_config.m_title;
+  auto firstInputBaseName    = QFileInfo{ m_config.m_firstInputFileName }.completeBaseName();
+  auto baseName              = !settings.m_mergeSetDestinationFromTitle ? firstInputBaseName
+                             : !m_config.m_title.isEmpty()              ? m_config.m_title
+                             :                                            firstInputBaseName;
   m_config.m_destinationAuto = generateUniqueOutputFileName(baseName, outputDir);
 
   ui->output->setText(m_config.m_destinationAuto);